update cmake template

+++ .tmpl/build_system/cmake_template/cb.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 echo Configuration and Build
-rm compile_commands.json
-rm build -rf
-mkdir build
+[ -f compile_commands.json ] && rm compile_commands.json
+[ -d build ] && rm -rf build
+mkdir -p build
 cd build ; cmake .. -DCMAKE_EXPORT_COMPILE_COMMANDS=YES -G Ninja \
   && ln -s build/compile_commands.json .. \
   && ninja -v -j14
